+++ paddle/phi/core/tensor_utils.h
@@ -14,6 +14,7 @@ limitations under the License. */
 
 #pragma once
 
+#include "paddle/fluid/framework/operator.h"
 #include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/core/device_context.h"
 #include "paddle/phi/core/selected_rows.h"
@@ -144,7 +145,36 @@ inline T GetValue(const Context& dev_ctx, const DenseTensor& x) {
   return value;
 }
 
-template <typename T>
-std::vector<T> GetDataFromTensor(const DenseTensor* x);
+template <typename T = int32_t>
+inline std::vector<T> GetDataFromTensor(const DenseTensor* x) {
+  std::vector<T> vec_new_data;
+  if (paddle::framework::TransToProtoVarType(x->dtype()) ==
+      paddle::framework::proto::VarType::INT32) {
+    auto* data = x->data<int>();
+    DenseTensor cpu_attr_tensor;
+    if (!paddle::platform::is_cpu_place(x->place())) {
+      paddle::framework::TensorCopySync(
+          *x, paddle::platform::CPUPlace(), &cpu_attr_tensor);
+      data = cpu_attr_tensor.data<int>();
+    }
+    vec_new_data = std::vector<T>(data, data + x->numel());
+  } else if (paddle::framework::TransToProtoVarType(x->dtype()) ==
+             paddle::framework::proto::VarType::INT64) {
+    auto* data = x->data<int64_t>();
+    DenseTensor cpu_attr_tensor;
+    if (!paddle::platform::is_cpu_place(x->place())) {
+      paddle::framework::TensorCopySync(
+          *x, paddle::platform::CPUPlace(), &cpu_attr_tensor);
+      data = cpu_attr_tensor.data<int64_t>();
+    }
+    // NOTE: Converting int64 to int32 may cause data overflow.
+    vec_new_data = std::vector<T>(data, data + x->numel());
+  } else {
+    PADDLE_THROW(paddle::platform::errors::InvalidArgument(
+        "The dtype of Tensor must be int32 or int64, but received: %s",
+        paddle::framework::TransToProtoVarType(x->dtype())));
+  }
+  return vec_new_data;
+}
 
 }  // namespace phi
